Wow that was a waste of time. Remember that WOO! Not so woo anymore
Modified the evaluator...twice...
Now it keeps track of how many times each person satisfies a,b,or c.
Once all is done the people who don't satisfy all of a variable are
eliminated from the vector passed in, resulting in the a b c vectors
only having the complete satisfiers.
Returns the number of conditions that can be satisfied in the current
conditions...unlike before where it only did if everything worked.