Currently doesn't allow for determining who is closest
If a person fits the preconditions it will work. If not it will evaluate
false.
Need to do a(nother) rewrite of the evaluation class to allow for
determining who was closest and which conditions were wrong.
Added a mandatory precondition to the story class for things that will
never create a storygoal