I don't like to preemptively woo
Added some sticky notes to the wall and came up with a(nother) "working"
system.
-added the condition class which stores information the condition
strings did before, but can also keep track of which people satisfy the
condition (game changer).
-rewrote the evaluator class, adding 3 new functions which simple look
at the people given and check if they satisfy the conditions. Only the
people who satisfy all the conditions are passed for the mandatory
conditions, and the people who satisfy the most conditions in the
optional are passed on. Another function was added which when passed a
condition simply checks it and returns true or false. This, for obvious
reasons, doesn't allow for variable inputs, but rather just works with
already set inputs (ex isPolice paul rather than isPolice a)

-It would appear author goals may be able to just become stories. More
investigation later.

-Once a story evaluate all of its conditions, the optional conditions
which the chosen people don't satisfy will need to be added as goals for
the director.