Untested and unsure if it even runs
-Worked on the main director loop
-Added the trigger condition
-Stories reside in the storyqueue at the beginning of each cycle. They
are each checked for completeness in terms of preconditions. If
complete, push them onto the working vector and check if their triggers
are also true. If so, then choose one and run it next time(The running
part is not complete).
-Goals are checked next. Every goal is compared to the current active
stories to see if any story is pursuing the goal. If so, move on, if not
find a story that will try to achieve the goal.
-Once a story is found, change the necessary variables associated with
the goal and test the mandatory conditions. If these pass push the story
onto the storyqueue and evaluate the rest of the options. If not, ditch
this story and move on to the next one(hopefully there will be one that
works)