Tested the director loop and it runs!
Still don't want to preemptively woo but things are looking promising.

Tested and debugged the director loop. Stories are being added to the
story queue and proper stories from there are being added to the working
list. Still not "running" and stories in the form.

Fixed some issues in the story and interpreter classes.

Currently adding a story will activate the director loop, and will run
until no changes are made.